@@ -2,8 +2,11 @@
 namespace Grav\Plugin;
 
 use Composer\Autoload\ClassLoader;
+use Grav\Common\Cache;
 use Grav\Common\Grav;
 use Grav\Common\Data;
+use Grav\Common\Language\Language;
+use Grav\Common\Page\Interfaces\PageInterface;
 use Grav\Common\Page\Page;
 use Grav\Common\Plugin;
 use Grav\Common\Uri;
@@ -11,13 +14,27 @@ use Grav\Common\Page\Pages;
 use Grav\Common\Utils;
 use Grav\Plugin\Sitemap\SitemapEntry;
 use RocketTheme\Toolbox\Event\Event;
+use Twig\TwigFunction;
 
 class SitemapPlugin extends Plugin
 {
     /**
      * @var array
      */
-    protected $sitemap = array();
+    protected $sitemap = false;
+    protected $route_data = [];
+
+    protected $multilang_skiplang_prefix = null;
+    protected $multilang_include_fallbacks = false;
+    protected $datetime_format = null;
+    protected $include_change_freq = true;
+    protected $default_change_freq = null;
+    protected $include_priority = true;
+    protected $default_priority = null;
+    protected $ignores = null;
+    protected $ignore_external = true;
+    protected $ignore_protected = true;
+    protected $ignore_redirect = true;
 
     /**
      * @return array
@@ -60,6 +77,7 @@ class SitemapPlugin extends Plugin
         if ($route && $route == $uri->path()) {
 
             $this->enable([
+                'onTwigInitialized' => ['onTwigInitialized', 0],
                 'onTwigTemplatePaths' => ['onTwigTemplatePaths', 0],
                 'onPagesInitialized' => ['onPagesInitialized', 0],
                 'onPageInitialized' => ['onPageInitialized', 0],
@@ -73,63 +91,70 @@ class SitemapPlugin extends Plugin
      */
     public function onPagesInitialized()
     {
-        // get grav instance and current language
-        $grav = Grav::instance();
-        $current_lang = $grav['language']->getLanguage() ?: 'en';
+        /** @var Cache $cache */
+        $cache = $this->grav['cache'];
 
         /** @var Pages $pages */
         $pages = $this->grav['pages'];
-        $routes = array_unique($pages->routes());
-        ksort($routes);
 
-        $ignores = (array) $this->config->get('plugins.sitemap.ignores');
-        $ignore_external = $this->config->get('plugins.sitemap.ignore_external');
-        $ignore_protected = $this->config->get('plugins.sitemap.ignore_protected');
+        $cache_id = md5('sitemap-data-'.$pages->getPagesCacheId());
+        $this->sitemap = $cache->fetch($cache_id);
 
-        foreach ($routes as $route => $path) {
-            $page = $pages->get($path);
-            $header = $page->header();
-            $external_url = $ignore_external ? isset($header->external_url) : false;
-            $protected_page = $ignore_protected ? isset($header->access) : false;
-            $page_ignored = $protected_page || $external_url || (isset($header->sitemap['ignore']) ? $header->sitemap['ignore'] : false);
-            $page_languages = $page->translatedLanguages();
-            $lang_available = (empty($page_languages) || array_key_exists($current_lang, $page_languages));
+        if ($this->sitemap === false) {
+            /** @var Language $language */
+            $language = $this->grav['language'];
+            $default_lang = $language->getDefault() ?: 'en';
+            $languages = $language->enabled() ? $language->getLanguages() : [$default_lang];
 
+            $this->multilang_skiplang_prefix = $this->config->get('system.languages.include_default_lang') ?  '' : $language->getDefault();
+            $this->multilang_include_fallbacks = $this->config->get('system.languages.pages_fallback_only') || !empty($this->config->get('system.languages.content_fallback'));
 
-            if ($page->published() && $page->routable() && !preg_match(sprintf("@^(%s)$@i", implode('|', $ignores)), $page->route()) && !$page_ignored && $lang_available ) {
-                
-                $entry = new SitemapEntry();
-                $entry->location = $page->canonical();
-                $entry->lastmod = date('Y-m-d', $page->modified());
+            $this->datetime_format = $this->config->get('plugins.sitemap.short_date_format') ? 'Y-m-d' : 'Y-m-d\TH:i:sP';
+            $this->include_change_freq = $this->config->get('plugins.sitemap.include_changefreq');
+            $this->default_change_freq = $this->config->get('plugins.sitemap.changefreq');
+            $this->include_priority = $this->config->get('plugins.sitemap.include_priority');
+            $this->default_priority = $this->config->get('plugins.sitemap.priority');
+            $this->ignores = (array) $this->config->get('plugins.sitemap.ignores');
+            $this->ignore_external = $this->config->get('plugins.sitemap.ignore_external');
+            $this->ignore_protected = $this->config->get('plugins.sitemap.ignore_protected');
+            $this->ignore_redirect = $this->config->get('plugins.sitemap.ignore_redirect');
 
-                // optional changefreq & priority that you can set in the page header
-                $entry->changefreq = (isset($header->sitemap['changefreq'])) ? $header->sitemap['changefreq'] : $this->config->get('plugins.sitemap.changefreq');
-                $entry->priority = (isset($header->sitemap['priority'])) ? $header->sitemap['priority'] : $this->config->get('plugins.sitemap.priority');
+            // Gather data
+            foreach ($languages as $lang) {
+                $language->init();
+                $language->setActive($lang);
+                $pages->reset();
+                $this->addRouteData($pages, $lang);
+            }
 
-                if (count($this->config->get('system.languages.supported', [])) > 0) {
-                    $entry->translated = $page->translatedLanguages(true);
-
-                    foreach($entry->translated as $lang => $page_route) {
-                        $page_route = $page->rawRoute();
-                        if ($page->home()) {
-                            $page_route = '';
+            // Build sitemap
+            foreach ($languages as $lang) {
+                foreach($this->route_data as $route => $route_data) {
+                    if ($data = $route_data[$lang] ?? null) {
+                        $entry = new SitemapEntry();
+                        $entry->setData($data);
+                        if ($language->enabled()) {
+                            foreach ($route_data as $l => $l_data) {
+                                $entry->addHreflangs(['hreflang' => $l, 'href' => $l_data['location']]);
+                                if ($l === $default_lang) {
+                                    $entry->addHreflangs(['hreflang' => 'x-default', 'href' => $l_data['location']]);
+                                }
+                            }
                         }
-
-                        $entry->translated[$lang] = $page_route;
+                        $this->sitemap[$data['route']] = $entry;
                     }
                 }
-
-                $this->sitemap[$route] = $entry;
             }
-        }
 
-        $additions = (array) $this->config->get('plugins.sitemap.additions');
-        foreach ($additions as $addition) {
-            if (isset($addition['location'])) {
-                $location = Utils::url($addition['location'], true);
-                $entry = new SitemapEntry($location,$addition['lastmod']??null,$addition['changefreq']??null, $addition['priority']??null);
-                $this->sitemap[$location] = $entry;
+            $additions = (array) $this->config->get('plugins.sitemap.additions');
+            foreach ($additions as $addition) {
+                if (isset($addition['location'])) {
+                    $location = Utils::url($addition['location'], true);
+                    $entry = new SitemapEntry($location,$addition['lastmod'] ?? null,$addition['changefreq'] ?? null, $addition['priority'] ?? null);
+                    $this->sitemap[$location] = $entry;
+                }
             }
+            $cache->save($cache_id, $this->sitemap);
         }
 
         $this->grav->fireEvent('onSitemapProcessed', new Event(['sitemap' => &$this->sitemap]));
@@ -141,15 +166,25 @@ class SitemapPlugin extends Plugin
         $route = $this->config->get('plugins.sitemap.route');
 
         if (is_null($page) || $page->route() !== $route) {
+            $extension = $this->grav['uri']->extension() ?? 'xml';
+
             // set a dummy page
             $page = new Page;
             $page->init(new \SplFileInfo(__DIR__ . '/pages/sitemap.md'));
+            $page->templateFormat($extension);
             unset($this->grav['page']);
             $this->grav['page'] = $page;
-
             $twig = $this->grav['twig'];
-            $twig->template = 'sitemap.xml.twig';
+            $twig->template = "sitemap.$extension.twig";
         }
+    }
+
+    // Access plugin events in this class
+    public function onTwigInitialized()
+    {
+        $this->grav['twig']->twig()->addFunction(
+            new TwigFunction('sort_sitemap_entries_by_language', [$this, 'sortSitemapEntriesByLanguage'])
+        );
     }
 
     /**
@@ -187,6 +222,81 @@ class SitemapPlugin extends Plugin
                 $extends = $blueprints->get('sitemap');
                 $blueprint->extend($extends, true);
                 $inEvent = false;
+            }
+        }
+    }
+
+    public function sortSitemapEntriesByLanguage()
+    {
+        $entries = [];
+
+        foreach ((array) $this->sitemap as $route => $entry) {
+            $lang = $entry->getLang();
+            unset($entry->hreflangs);
+            unset($entry->image);
+            if ($lang === null) {
+                $lang = $this->grav['language']->getDefault() ?: 'en';
+            }
+            $entries[$lang][$route] = $entry;
+        }
+        return $entries;
+    }
+
+    protected function addRouteData($pages, $lang)
+    {
+        $routes = array_unique($pages->routes());
+        ksort($routes);
+
+        foreach ($routes as $route => $path) {
+            /** @var PageInterface $page */
+            $page = $pages->get($path);
+
+            $header = $page->header();
+            $external_url = $this->ignore_external ? isset($header->external_url) : false;
+            $protected_page = $this->ignore_protected ? isset($header->access) : false;
+            $redirect_page = $this->ignore_redirect ? isset($header->redirect) : false;
+            $config_ignored = preg_match(sprintf("@^(%s)$@i", implode('|', $this->ignores)), $page->route());
+            $page_ignored = $protected_page || $external_url || $redirect_page || (isset($header->sitemap['ignore']) ? $header->sitemap['ignore'] : false);
+
+            if ($page->routable() && $page->published() && !$config_ignored && !$page_ignored) {
+                $page_languages = array_keys($page->translatedLanguages());
+                $include_lang = $this->multilang_skiplang_prefix !== $lang;
+                $location = $page->canonical($include_lang);
+                $page_route = $page->url(false, $include_lang);
+
+                $lang_route = [
+                    'title' => $page->title(),
+                    'route' => $page_route,
+                    'lang' => $lang,
+                    'translated' => in_array($lang, $page_languages),
+                    'location' => $location,
+                    'lastmod' => date($this->datetime_format, $page->modified()),
+                ];
+
+                if ($this->include_change_freq) {
+                    $lang_route['changefreq'] = $header->sitemap['changefreq'] ?? $this->default_change_freq;
+                }
+                if ($this->include_priority) {
+                    $lang_route['priority']  = $header->sitemap['priority'] ?? $this->default_priority;
+                }
+
+                // optional add image
+                $images = $header->sitemap['images'] ?? $this->config->get('plugins.sitemap.images') ?? [];
+
+                if (isset($images)) {
+                    foreach ($images as $image => $values) {
+                        if (isset($values['loc'])) {
+                            $images[$image]['loc'] = $page->media()[$values['loc']]->url();
+                        } else {
+                            unset($images[$image]);
+                        }
+                    }
+                    $lang_route['images'] = $images;
+                }
+
+
+
+                $this->route_data[$route][$lang] = $lang_route;
             }
         }
     }
