@@ -414,12 +414,11 @@ class Email
 
         $config = $grav['config']->get('plugins.email.queue');
 
-        $queue = static::getQueue();
-        $spool = $queue->getSpool();
-        $spool->setMessageLimit($config['flush_msg_limit']);
-        $spool->setTimeLimit($config['flush_time_limit']);
-
         try {
+            $queue = static::getQueue();
+            $spool = $queue->getSpool();
+            $spool->setMessageLimit($config['flush_msg_limit']);
+            $spool->setTimeLimit($config['flush_time_limit']);
             $failures = [];
             $result = $spool->flushQueue(static::getTransport(), $failures);
             return $result . ' messages flushed from queue...';
