@@ -13,31 +13,54 @@ class Forms
     /** @var FormInterface|null */
     private $form;
 
+    /**
+     * Forms constructor.
+     */
     public function __construct()
     {
         $this->registerType('form', new FormFactory());
     }
 
+    /**
+     * @param string $type
+     * @param FormFactoryInterface $factory
+     */
     public function registerType(string $type, FormFactoryInterface $factory): void
     {
         $this->types[$type] = $factory;
     }
 
+    /**
+     * @param string $type
+     */
     public function unregisterType($type): void
     {
         unset($this->types[$type]);
     }
 
+    /**
+     * @param string $type
+     * @return bool
+     */
     public function hasType(string $type): bool
     {
         return isset($this->types[$type]);
     }
 
+    /**
+     * @return array
+     */
     public function getTypes(): array
     {
         return array_keys($this->types);
     }
 
+    /**
+     * @param PageInterface $page
+     * @param string|null $name
+     * @param array|null $form
+     * @return FormInterface|null
+     */
     public function createPageForm(PageInterface $page, string $name = null, array $form = null): ?FormInterface
     {
         if (null === $form) {
@@ -52,11 +75,12 @@ class Forms
         $factory = $this->types[$type] ?? null;
 
         if ($factory) {
-            if (method_exists($factory, 'createFormForPage')) {
+            if (is_callable([$factory, 'createFormForPage'])) {
                 return $factory->createFormForPage($page, $name, $form);
             }
 
             if ($page instanceof Page) {
+                // @phpstan-ignore-next-line
                 return $factory->createPageForm($page, $name, $form);
             }
         }
@@ -64,16 +88,28 @@ class Forms
         return null;
     }
 
+    /**
+     * @return FormInterface|null
+     */
     public function getActiveForm(): ?FormInterface
     {
         return $this->form;
     }
 
+    /**
+     * @param FormInterface $form
+     * @return void
+     */
     public function setActiveForm(FormInterface $form): void
     {
         $this->form = $form;
     }
 
+    /**
+     * @param PageInterface $page
+     * @param string|null $name
+     * @return array
+     */
     protected function getPageParameters(PageInterface $page, ?string $name): array
     {
         $forms = $page->forms();
